Move ComponentModel classes into new project
