+++ ReactiveComponentModel/DelegateRule.cs
