+++ ReactiveComponentModel/EditableObject.cs
