+++ ReactiveComponentModel/EventAggregator.cs
