+++ ReactiveComponentModel/IEventAggregator.cs
