+++ ReactiveComponentModel/NotifyPropertyChanges.cs
@@ -1,0 +1,244 @@
+namespace Common
+{
+    using System;
+    using System.ComponentModel;
+    using System.Diagnostics;
+    using System.Reactive.Linq;
+    using System.Reflection;
+    using System.Runtime.CompilerServices;
+
+    /// <summary>
+    /// Notifies subscribers that a property in this instance is changing or has changed.
+    /// </summary>
+    public abstract class NotifyPropertyChanges : Disposable, INotifyPropertyChanged, INotifyPropertyChanging
+    {
+        /// <summary>
+        /// Occurs after a property value changes.
+        /// </summary>
+        event PropertyChangedEventHandler INotifyPropertyChanged.PropertyChanged
+        {
+            add { this.propertyChanged += value; }
+            remove { this.propertyChanged -= value; }
+        }
+
+        /// <summary>
+        /// Occurs before a property value is changed.
+        /// </summary>
+        event PropertyChangingEventHandler INotifyPropertyChanging.PropertyChanging
+        {
+            add { this.propertyChanging += value; }
+            remove { this.propertyChanging -= value; }
+        }
+
+#pragma warning disable IDE1006 // Naming Styles
+        private event PropertyChangedEventHandler propertyChanged;
+        private event PropertyChangingEventHandler propertyChanging;
+#pragma warning restore IDE1006 // Naming Styles
+
+        /// <summary>
+        /// Gets the when property changed observable event. Occurs when a property value changes.
+        /// </summary>
+        /// <value>
+        /// The when property changed observable event.
+        /// </value>
+        public IObservable<string> WhenPropertyChanged
+        {
+            get
+            {
+                this.ThrowIfDisposed();
+
+                return Observable
+                    .FromEventPattern<PropertyChangedEventHandler, PropertyChangedEventArgs>(
+                        h => this.propertyChanged += h,
+                        h => this.propertyChanged -= h)
+                    .Select(x => x.EventArgs.PropertyName);
+            }
+        }
+
+        /// <summary>
+        /// Gets the when property changing observable event. Occurs when a property value is changing.
+        /// </summary>
+        /// <value>
+        /// The when property changing observable event.
+        /// </value>
+        public IObservable<string> WhenPropertyChanging
+        {
+            get
+            {
+                this.ThrowIfDisposed();
+
+                return Observable
+                    .FromEventPattern<PropertyChangingEventHandler, PropertyChangingEventArgs>(
+                        h => this.propertyChanging += h,
+                        h => this.propertyChanging -= h)
+                    .Select(x => x.EventArgs.PropertyName);
+            }
+        }
+
+        /// <summary>
+        /// Raises the PropertyChanged event.
+        /// </summary>
+        /// <param name="propertyName">Name of the property.</param>
+        protected virtual void OnPropertyChanged([CallerMemberName] string propertyName = null)
+        {
+            Debug.Assert(
+                string.IsNullOrEmpty(propertyName) ||
+                (this.GetType().GetRuntimeProperty(propertyName) != null),
+                "Check that the property name exists for this instance.");
+            this.propertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
+        }
+
+        /// <summary>
+        /// Raises the PropertyChanged event.
+        /// </summary>
+        /// <param name="propertyNames">The property names.</param>
+        protected void OnPropertyChanged(params string[] propertyNames)
+        {
+            if (propertyNames == null)
+            {
+                throw new ArgumentNullException(nameof(propertyNames));
+            }
+
+            foreach (var propertyName in propertyNames)
+            {
+                this.OnPropertyChanged(propertyName);
+            }
+        }
+
+        /// <summary>
+        /// Raises the PropertyChanging event.
+        /// </summary>
+        /// <param name="propertyName">Name of the property.</param>
+        protected virtual void OnPropertyChanging([CallerMemberName] string propertyName = null)
+        {
+            Debug.Assert(
+                string.IsNullOrEmpty(propertyName) ||
+                (this.GetType().GetRuntimeProperty(propertyName) != null),
+                "Check that the property name exists for this instance.");
+            this.propertyChanging?.Invoke(this, new PropertyChangingEventArgs(propertyName));
+        }
+
+        /// <summary>
+        /// Raises the PropertyChanging event.
+        /// </summary>
+        /// <param name="propertyNames">The property names.</param>
+        protected void OnPropertyChanging(params string[] propertyNames)
+        {
+            if (propertyNames == null)
+            {
+                throw new ArgumentNullException(nameof(propertyNames));
+            }
+
+            foreach (var propertyName in propertyNames)
+            {
+                this.OnPropertyChanging(propertyName);
+            }
+        }
+
+        /// <summary>
+        /// Sets the value of the property to the specified value if it has changed.
+        /// </summary>
+        /// <typeparam name="TProp">The type of the property.</typeparam>
+        /// <param name="currentValue">The current value of the property.</param>
+        /// <param name="newValue">The new value of the property.</param>
+        /// <param name="propertyName">Name of the property.</param>
+        /// <returns><c>true</c> if the property was changed, otherwise <c>false</c>.</returns>
+        protected bool SetProperty<TProp>(
+            ref TProp currentValue,
+            TProp newValue,
+            [CallerMemberName] string propertyName = null)
+        {
+            this.ThrowIfDisposed();
+
+            if (!object.Equals(currentValue, newValue))
+            {
+                this.OnPropertyChanging(propertyName);
+                currentValue = newValue;
+                this.OnPropertyChanged(propertyName);
+
+                return true;
+            }
+
+            return false;
+        }
+
+        /// <summary>
+        /// Sets the value of the property to the specified value if it has changed.
+        /// </summary>
+        /// <typeparam name="TProp">The type of the property.</typeparam>
+        /// <param name="currentValue">The current value of the property.</param>
+        /// <param name="newValue">The new value of the property.</param>
+        /// <param name="propertyNames">The names of all properties changed.</param>
+        /// <returns><c>true</c> if the property was changed, otherwise <c>false</c>.</returns>
+        protected bool SetProperty<TProp>(
+            ref TProp currentValue,
+            TProp newValue,
+            params string[] propertyNames)
+        {
+            this.ThrowIfDisposed();
+
+            if (!object.Equals(currentValue, newValue))
+            {
+                this.OnPropertyChanging(propertyNames);
+                currentValue = newValue;
+                this.OnPropertyChanged(propertyNames);
+
+                return true;
+            }
+
+            return false;
+        }
+
+        /// <summary>
+        /// Sets the value of the property to the specified value if it has changed.
+        /// </summary>
+        /// <param name="equal">A function which returns <c>true</c> if the property value has changed, otherwise <c>false</c>.</param>
+        /// <param name="action">The action where the property is set.</param>
+        /// <param name="propertyName">Name of the property.</param>
+        /// <returns><c>true</c> if the property was changed, otherwise <c>false</c>.</returns>
+        protected bool SetProperty(
+            Func<bool> equal,
+            Action action,
+            [CallerMemberName] string propertyName = null)
+        {
+            this.ThrowIfDisposed();
+
+            if (equal())
+            {
+                return false;
+            }
+
+            this.OnPropertyChanging(propertyName);
+            action();
+            this.OnPropertyChanged(propertyName);
+
+            return true;
+        }
+
+        /// <summary>
+        /// Sets the value of the property to the specified value if it has changed.
+        /// </summary>
+        /// <param name="equal">A function which returns <c>true</c> if the property value has changed, otherwise <c>false</c>.</param>
+        /// <param name="action">The action where the property is set.</param>
+        /// <param name="propertyNames">The property names.</param>
+        /// <returns><c>true</c> if the property was changed, otherwise <c>false</c>.</returns>
+        protected bool SetProperty(
+            Func<bool> equal,
+            Action action,
+            params string[] propertyNames)
+        {
+            this.ThrowIfDisposed();
+
+            if (equal())
+            {
+                return false;
+            }
+
+            this.OnPropertyChanging(propertyNames);
+            action();
+            this.OnPropertyChanged(propertyNames);
+
+            return true;
+        }
+    }
+}
