+++ ReactiveComponentModel/RevertibleChangeTracking.cs
