+++ ReactiveComponentModel/Rule.cs
