+++ ReactiveComponentModel/RuleCollection.cs
